+++ kf_lib/happenings/story/__init__.py
@@ -13,8 +13,10 @@ from ._jade_table import JadeTableStory
 from ._ninja_turtles import NinjaTurtlesStory
 from ._renowned_master import RenownedMasterStory
 from ._saints_day_riot import SaintsDayRiotStory
+from ._school_attack import SchoolAttackStory
 from ._stolen_treasures import StolenTreasuresStory
 from ._strange_dreams import StrangeDreamsStory
+from ._student_rivalry import StudentRivalryStory
 from ._wrong_pouch import WrongPouchStory